requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
requests
python-dotenv
openai
beautifulsoup4
lxml
transformers
torch
sentencepiece
accelerate
bitsandbytes
